added RandomForest fixed some warnings

+++ modules/features/include/v4r/features/sift_local_estimator.h
@@ -382,7 +382,7 @@ public:
             int u,v;
             v = (int)(ks[i].y+.5);
             u = (int)(ks[i].x+.5);
-            if(u >= 0 && v >= 0 && u < mask_cloud.width && v < mask_cloud.height && mask_cloud.at(u,v))
+            if(u >= 0 && v >= 0 && u < (int)mask_cloud.width && v < (int)mask_cloud.height && mask_cloud.at(u,v))
             {
                 if(pcl_isfinite(in->at(u,v).z) && pcl_isfinite(in->at(u,v).x) && pcl_isfinite(in->at(u,v).y))
                 {
